Allow creating hard line breaks on mobile devices + documentation
This allows to create hard line breaks on mobile, currently a hard
line break could only created using `shift + enter`, but most mobile
keyboards do not support the `shift` modifier.
So on mobile devices pressing the enter key will trigger a hard line break
and pressing it again will convert it to a new paragraph.

On desktop devices the behavior is unchanged, but documentation on how to
add hard line breaks is added.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>