fix: load initial content from session if available
If an editor session is already available
tiptap will read the content from the ydoc.
So adding initial content will end up showing twice.

However if no session is available we still need to load initial content.
It will show up in the editor as created by the first step.

Signed-off-by: Max <max@nextcloud.com>